server/requirements.txt
flask
gunicorn
requests
beautifulsoup4
nltk
newspaper3k
lxml_html_clean
flask_cors
pymongo
datetime
schedule